feat: validate token endpoint checks if token is valid and returns the user
