--- src/main/java/dtos/UserDTO.java
@@ -1,116 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package dtos;
-
-import entities.User;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-/**
- *
- * @author peter
- */
-public class UserDTO {
-    
-    
-    private String userName;
-    private String password;
-    private List<RoleDTO> roles = new ArrayList<>();
-
-    public UserDTO(String userName, String password) {
-       
-        this.userName = userName;
-        this.password = password;
-       
-    }
-
-    public UserDTO(User user) {
-        this.userName = user.getUserName();
-        this.password = user.getUserPass();
-        this.roles = RoleDTO.getAllRoleDtoes(user.getRoleList());
-    }
-    
-    public List<String> getRolesAsStrings() {
-        if (roles.isEmpty()) {
-            return null;
-        }
-        List<String> rolesAsStrings = new ArrayList<>();
-        roles.forEach((role) -> {
-            rolesAsStrings.add(role.getRoleName());
-        });
-        return rolesAsStrings;
-    }
-
-       public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public List<RoleDTO> getRoles() {
-        return roles;
-    }
-
-    public void setRoles(List<RoleDTO> roles) {
-        this.roles = roles;
-    }
-
-    @Override
-    public int hashCode() {
-        int hash = 7;
-        hash = 53 * hash + Objects.hashCode(this.userName);
-        hash = 53 * hash + Objects.hashCode(this.password);
-        hash = 53 * hash + Objects.hashCode(this.roles);
-        return hash;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        final UserDTO other = (UserDTO) obj;
-        if (!Objects.equals(this.userName, other.userName)) {
-            return false;
-        }
-        if (!Objects.equals(this.password, other.password)) {
-            return false;
-        }
-      
-        if (!Objects.equals(this.roles, other.roles)) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "UserDTO{" + "userName=" + userName + ", hashPassword=" + password + '}';
-    }
-    
-    
-
-    
-    
-    
-}
