+++ src/main/java/rest/UserResource.java
@@ -7,18 +7,12 @@ package rest;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import dtos.UserDTO;
+import dtos.user.UserDTO;
 import entities.UserRepository;
 import facades.UserFacade;
-import javax.persistence.EntityManagerFactory;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.Produces;
-import javax.ws.rs.GET;
 import javax.ws.rs.Path;
-import javax.ws.rs.PUT;
-import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import rest.provider.Provider;
