+++ src/test/java/rest/UserResourceTest.java
@@ -5,20 +5,16 @@
  */
 package rest;
 
-import dtos.UserDTO;
+import dtos.user.UserDTO;
 import entities.Role;
 import static io.restassured.RestAssured.given;
 import io.restassured.http.ContentType;
 import javax.persistence.EntityManager;
 import org.glassfish.grizzly.http.util.HttpStatus;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import static rest.SetupRestTests.emf;
-import static rest.SetupRestTests.setupServer;
-import static rest.SetupRestTests.shutdownServer;
 
 
 /**
@@ -43,14 +39,15 @@ public class UserResourceTest extends SetupRestTests {
         Role role = new Role("user");
         try {
             em.getTransaction().begin();
+            em.createQuery("DELETE FROM BasketItem").executeUpdate();
+            em.createQuery("DELETE FROM Basket").executeUpdate();
             em.createQuery("DELETE FROM Role").executeUpdate();
-            em.createQuery("DELETE FROM User").executeUpdate();
+            em.createNamedQuery("User.deleteAllRows").executeUpdate();
             em.persist(role);
             em.getTransaction().commit();
-        }  finally {       
+        }  finally {
+            em.close();
         }
-        em.close();
-        
     }
     
     @Test
